+++ Microsoft.Azure.Cosmos/src/FeedIteratorCore.cs
@@ -8,102 +8,42 @@ namespace Microsoft.Azure.Cosmos
     using System.Collections.Generic;
     using System.IO;
     using System.Net;
-    using System.Text;
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Azure.Cosmos.CosmosElements;
     using Microsoft.Azure.Cosmos.Query.Core;
-    using Microsoft.Azure.Cosmos.Query.Core.Monads;
-    using Microsoft.Azure.Cosmos.Resource.CosmosExceptions;
     using Microsoft.Azure.Documents;
     using static Microsoft.Azure.Documents.RuntimeConstants;
 
     /// <summary>
-    /// Cosmos feed stream iterator. This is used to get the query responses with a Stream content
+    /// Cosmos feed stream iterator. This is used to get the query responses with a Stream content for non-partitioned results
     /// </summary>
     internal sealed class FeedIteratorCore : FeedIteratorInternal
     {
-        private readonly ContainerCore containerCore;
         private readonly CosmosClientContext clientContext;
         private readonly Uri resourceLink;
         private readonly ResourceType resourceType;
         private readonly SqlQuerySpec querySpec;
         private bool hasMoreResultsInternal;
-        private FeedTokenInternal feedTokenInternal;
 
-        internal static FeedIteratorCore CreateForNonPartitionedResource( 
+        public FeedIteratorCore(
             CosmosClientContext clientContext,
             Uri resourceLink,
             ResourceType resourceType,
             QueryDefinition queryDefinition,
             string continuationToken,
-            QueryRequestOptions options)
-        {
-            return new FeedIteratorCore(
-                clientContext: clientContext,
-                containerCore: null,
-                resourceLink: resourceLink,
-                resourceType: resourceType,
-                queryDefinition: queryDefinition,
-                continuationToken: continuationToken,
-                feedTokenInternal: null,
-                options: options);
-        }
-
-        internal static FeedIteratorCore CreateForPartitionedResource(
-            ContainerCore containerCore,
-            Uri resourceLink,
-            ResourceType resourceType,
-            QueryDefinition queryDefinition,
-            string continuationToken,
-            FeedTokenInternal feedTokenInternal,
-            QueryRequestOptions options)
-        {
-            if (containerCore == null)
-            {
-                throw new ArgumentNullException(nameof(containerCore));
-            }
-
-            return new FeedIteratorCore(
-                containerCore: containerCore,
-                clientContext: containerCore.ClientContext,
-                resourceLink: resourceLink,
-                resourceType: resourceType,
-                queryDefinition: queryDefinition,
-                continuationToken: continuationToken,
-                feedTokenInternal: feedTokenInternal,
-                options: options);
-        }
-
-        private FeedIteratorCore(
-            ContainerCore containerCore,
-            CosmosClientContext clientContext,
-            Uri resourceLink,
-            ResourceType resourceType,
-            QueryDefinition queryDefinition,
-            string continuationToken,
-            FeedTokenInternal feedTokenInternal,
             QueryRequestOptions options)
         {
             this.resourceLink = resourceLink;
-            this.containerCore = containerCore;
             this.clientContext = clientContext;
             this.resourceType = resourceType;
             this.querySpec = queryDefinition?.ToSqlQuerySpec();
-            this.feedTokenInternal = feedTokenInternal;
-            this.ContinuationToken = continuationToken ?? this.feedTokenInternal?.GetContinuation();
+            this.ContinuationToken = continuationToken;
             this.requestOptions = options;
             this.hasMoreResultsInternal = true;
         }
 
         public override bool HasMoreResults => this.hasMoreResultsInternal;
-
-#if PREVIEW
-        public override
-#else
-        internal
-#endif
-        FeedToken FeedToken => this.feedTokenInternal;
 
         /// <summary>
         /// The query options for the result set
@@ -143,25 +83,6 @@ namespace Microsoft.Azure.Cosmos
                 operation = OperationType.Query;
             }
 
-            if (this.feedTokenInternal == null)
-            {
-                TryCatch<FeedTokenInternal> tryCatchFeedTokeninternal = await this.TryInitializeFeedTokenAsync(cancellationToken);
-                if (!tryCatchFeedTokeninternal.Succeeded)
-                {
-                    if (tryCatchFeedTokeninternal.Exception.InnerException is CosmosException cosmosException)
-                    {
-                        return cosmosException.ToCosmosResponseMessage(new RequestMessage(method: null, requestUri: null, diagnosticsContext: diagnostics));
-                    }
-
-                    return CosmosExceptionFactory.CreateInternalServerErrorException(
-                        message: tryCatchFeedTokeninternal.Exception.InnerException.Message,
-                        innerException: tryCatchFeedTokeninternal.Exception.InnerException,
-                        diagnosticsContext: diagnostics).ToCosmosResponseMessage(new RequestMessage(method: null, requestUri: null, diagnosticsContext: diagnostics));
-                }
-
-                this.feedTokenInternal = tryCatchFeedTokeninternal.Result;
-            }
-
             ResponseMessage response = await this.clientContext.ProcessResourceOperationStreamAsync(
                resourceUri: this.resourceLink,
                resourceType: this.resourceType,
@@ -178,25 +99,12 @@ namespace Microsoft.Azure.Cosmos
                        request.Headers.Add(HttpConstants.HttpHeaders.ContentType, MediaTypes.QueryJson);
                        request.Headers.Add(HttpConstants.HttpHeaders.IsQuery, bool.TrueString);
                    }
-
-                   this.feedTokenInternal?.EnrichRequest(request);
                },
                diagnosticsContext: diagnostics,
                cancellationToken: cancellationToken);
 
-            // Retry in case of splits or other scenarios only on partitioned resources
-            if (this.containerCore != null
-                && await this.feedTokenInternal.ShouldRetryAsync(this.containerCore, response, cancellationToken))
-            {
-                return await this.ReadNextInternalAsync(diagnostics, cancellationToken);
-            }
-
             if (response.IsSuccessStatusCode)
             {
-                this.feedTokenInternal.UpdateContinuation(response.Headers.ContinuationToken);
-                this.ContinuationToken = this.feedTokenInternal.GetContinuation();
-                this.hasMoreResultsInternal = !this.feedTokenInternal.IsDone;
-
                 if (this.clientContext.ClientOptions?.Encryptor != null && response.Content != null)
                 {
                     CosmosArray cosmosArray = CosmosElementSerializer.ToCosmosElements(
@@ -214,44 +122,10 @@ namespace Microsoft.Azure.Cosmos
                         decryptionInfo);
                 }
             }
-            else
-            {
-                this.hasMoreResultsInternal = false;
-            }
 
+            this.ContinuationToken = response.Headers.ContinuationToken;
+            this.hasMoreResultsInternal = this.ContinuationToken != null && response.StatusCode != HttpStatusCode.NotModified;
             return response;
-        }
-
-        private async Task<TryCatch<FeedTokenInternal>> TryInitializeFeedTokenAsync(CancellationToken cancellationToken)
-        {
-            string containerRId = string.Empty;
-            if (this.containerCore != null)
-            {
-                try
-                {
-                    containerRId = await this.containerCore.GetRIDAsync(cancellationToken);
-                }
-                catch (Exception cosmosException)
-                {
-                    return TryCatch<FeedTokenInternal>.FromException(cosmosException);
-                }
-            }
-
-            // Create FeedToken for the full Range
-            FeedTokenEPKRange feedTokenInternal = new FeedTokenEPKRange(
-                containerRId,
-                new PartitionKeyRange()
-                {
-                    MinInclusive = Documents.Routing.PartitionKeyInternal.MinimumInclusiveEffectivePartitionKey,
-                    MaxExclusive = Documents.Routing.PartitionKeyInternal.MaximumExclusiveEffectivePartitionKey
-                });
-            // Initialize with the ContinuationToken that the user passed, if any
-            if (this.ContinuationToken != null)
-            {
-                feedTokenInternal.UpdateContinuation(this.ContinuationToken);
-            }
-
-            return TryCatch<FeedTokenInternal>.FromResult(feedTokenInternal);
         }
 
         public override CosmosElement GetCosmsoElementContinuationToken()
@@ -283,10 +157,6 @@ namespace Microsoft.Azure.Cosmos
         {
             return this.feedIterator.GetCosmsoElementContinuationToken();
         }
-
-#if PREVIEW
-        public override FeedToken FeedToken => this.feedIterator.FeedToken;
-#endif
 
         /// <summary>
         /// Get the next set of results from the cosmos service