+++ Microsoft.Azure.Cosmos/src/Resource/QueryResponses/ChangeFeedPartitionKeyResultSetIteratorCore.cs
@@ -20,7 +20,6 @@ namespace Microsoft.Azure.Cosmos
         private readonly CosmosClientContext clientContext;
         private readonly ContainerCore container;
         private readonly ChangeFeedRequestOptions changeFeedOptions;
-        private readonly FeedTokenInternal feedToken;
         private string continuationToken;
         private string partitionKeyRangeId;
         private bool hasMoreResultsInternal;
@@ -49,8 +48,6 @@ namespace Microsoft.Azure.Cosmos
             this.MaxItemCount = maxItemCount;
             this.continuationToken = continuationToken;
             this.partitionKeyRangeId = partitionKeyRangeId;
-            this.feedToken = new FeedTokenPartitionKeyRange(this.partitionKeyRangeId);
-            this.feedToken.UpdateContinuation(this.continuationToken);
         }
 
         /// <summary>
@@ -64,13 +61,6 @@ namespace Microsoft.Azure.Cosmos
         {
             throw new NotImplementedException();
         }
-
-#if PREVIEW
-        public override
-#else
-        internal
-#endif
-        FeedToken FeedToken => this.feedToken;
 
         /// <summary>
         /// Get the next set of results from the cosmos service
@@ -87,7 +77,6 @@ namespace Microsoft.Azure.Cosmos
                     ResponseMessage response = task.Result;
                     // Change Feed uses ETAG
                     this.continuationToken = response.Headers.ETag;
-                    this.feedToken.UpdateContinuation(this.continuationToken);
                     this.hasMoreResultsInternal = response.StatusCode != HttpStatusCode.NotModified;
                     response.Headers.ContinuationToken = this.continuationToken;
                     return response;